v1.0.2 update - Now installable via PIPX, default data dirctory
