+++ pyproject.toml
@@ -1,0 +1,131 @@
+[project]
+name = "mitre-attack-mcp"
+version = "1.0.2"
+description = "A Model-Context Protocol server for the MITRE ATT&CK knowledge base"
+authors = [
+    {name = "stoyky"}
+]
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+"annotated-types==0.7.0",
+"antlr4-python3-runtime==4.9.3",
+"anyio==4.9.0",
+"arrow==1.3.0",
+"attrs==25.3.0",
+"black==25.1.0",
+"certifi==2025.1.31",
+"charset-normalizer==3.4.1",
+"check-wheel-contents==0.6.1",
+"click==8.1.8",
+"colorama==0.4.6",
+"colour==0.1.5",
+"coverage==7.8.0",
+"cpe==1.3.1",
+"cybox==2.1.0.21",
+"deepdiff==8.4.2",
+"drawsvg==2.3.0",
+"et_xmlfile==2.0.0",
+"exceptiongroup==1.2.2",
+"flake8==7.2.0",
+"flake8-docstrings==1.7.0",
+"fqdn==1.5.1",
+"h11==0.14.0",
+"httpcore==1.0.7",
+"httpx==0.28.1",
+"httpx-sse==0.4.0",
+"idna==3.10",
+"iniconfig==2.1.0",
+"isoduration==20.11.0",
+"isort==6.0.1",
+"jsonpointer==3.0.0",
+"jsonschema==4.23.0",
+"jsonschema-specifications==2024.10.1",
+"loguru==0.7.3",
+"lxml==5.3.1",
+"maec==4.1.0.17",
+"Markdown==3.7",
+"markdown-it-py==3.0.0",
+"mccabe==0.7.0",
+"mcp==1.6.0",
+"mdurl==0.1.2",
+"mitreattack-python==4.0.0",
+"mixbox==1.0.5",
+"mypy-extensions==1.0.0",
+"netaddr==1.3.0",
+"numpy==2.2.4",
+"openpyxl==3.1.5",
+"ordered-set==4.1.0",
+"orderly-set==5.3.0",
+"packaging==24.2",
+"pandas==2.2.3",
+"pathspec==0.12.1",
+"pillow==11.1.0",
+"platformdirs==4.3.7",
+"pluggy==1.5.0",
+"pluralizer==1.2.0",
+"pooch==1.8.2",
+"pycodestyle==2.13.0",
+"pycountry==24.6.1",
+"pydantic==2.11.1",
+"pydantic-settings==2.8.1",
+"pydantic_core==2.33.0",
+"pydocstyle==6.3.0",
+"pyflakes==3.3.1",
+"Pygments==2.19.1",
+"pyinstrument==5.0.1",
+"pytest==8.3.5",
+"pytest-cov==6.0.0",
+"pytest-dotenv==0.5.2",
+"python-dateutil==2.9.0.post0",
+"python-dotenv==1.1.0",
+"pytz==2025.2",
+"referencing==0.36.2",
+"requests==2.32.3",
+"rfc3339-validator==0.1.4",
+"rfc3986-validator==0.1.1",
+"rich==14.0.0",
+"rpds-py==0.24.0",
+"ruff==0.11.2",
+"setuptools==80.1.0",
+"shellingham==1.5.4",
+"simplejson==3.20.1",
+"six==1.17.0",
+"sniffio==1.3.1",
+"snowballstemmer==2.2.0",
+"sse-starlette==2.2.1",
+"starlette==0.46.1",
+"stix==1.2.0.11",
+"stix2==3.0.1",
+"stix2-elevator==4.1.7",
+"stix2-patterns==2.0.0",
+"stix2-validator==3.2.0",
+"stixmarx==1.0.8",
+"tabulate==0.9.0",
+"taxii2-client==2.3.0",
+"tomli==2.2.1",
+"tqdm==4.67.1",
+"typer==0.15.2",
+"types-python-dateutil==2.9.0.20241206",
+"typing-inspection==0.4.0",
+"typing_extensions==4.13.0",
+"tzdata==2025.2",
+"uri-template==1.3.0",
+"urllib3==2.3.0",
+"uvicorn==0.34.0",
+"weakrefmethod==1.0.3",
+"webcolors==24.11.1",
+"wheel-filename==1.4.2",
+"win32_setctime==1.2.0",
+"XlsxWriter==3.2.2"
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
+
+
+[project.scripts]
+mitre-attack-mcp = "mitre_attack_mcp.server:main"
+
